End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -61,12 +62,14 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
 
-            auction.Seller = "test";
+            auction.Seller = User.Identity?.Name ?? "Unknown Seller";
 
             _context.Auctions.Add(auction);
 
@@ -83,6 +86,8 @@ namespace AuctionService.Controllers
             return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
         }
 
+
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction([FromRoute] Guid id, [FromBody] UpdateAuctionDto updateAuctionDto)
         {
@@ -93,6 +98,11 @@ namespace AuctionService.Controllers
             if (auction == null)
             {
                 return NotFound();
+            }
+
+            if (auction.Seller != User.Identity?.Name)
+            {
+                return Forbid();
             }
 
 
@@ -113,6 +123,7 @@ namespace AuctionService.Controllers
             return BadRequest("Problem saving changes");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction([FromRoute] Guid id)
         {
@@ -122,6 +133,10 @@ namespace AuctionService.Controllers
             if (auction == null)
             {
                 return NotFound();
+            }
+            if (auction.Seller != User.Identity?.Name)
+            {
+                return Forbid();
             }
 
             _context.Auctions.Remove(auction);
